+++ changelog.md
@@ -21,7 +21,7 @@
 - new folder ./dropins/designs for custom design CSS files (please adjust urls of e.g. background images in the files)
 - page and service for backup and restore of configuration, dropins and custom pages (thanks to @hijacker7)
 - new Italian language package (granzie tante a @hijacker7)
-- eibd/knxd driver now supports status and control GAs in one item, separated by a colon. Format is "GA_state:GA_control", e.g. "1/0/0:1/0/1"
+- smarthomeNG driver and eibd/knxd driver now support status and control items/GAs in one item, separated by a colon. Format is "item_status:item_control" (or e.g. "1/0/0:1/0/1" for eibd/knxd)
 
 ### Improvements
 - Template checker now ignores multi-line comments